refactor(agent): drop dead nil checks and move shared defaults
agent.NewManager never returns an error; it had a single
return m, nil path. Drop the error from the signature so the
agent.go call site does not pretend to handle a failure that
cannot happen. Remove the five dead 'if a.contextManager != nil'
guards that followed.

Move the hardcoded default scan-root and allow-root lists out
of agent/agentcontext_seed.go and into the agentcontext package
itself as DefaultBuiltinRoots() and DefaultAllowedRoots(). The
package owns the policy now; the agent passes them in
explicitly, matching the pattern other defaults (Clock,
Debounce, SchemaVersion) follow.

agentcontext_seed.go is gone. The one helper that legitimately
lives in package agent, initialContextSources (which bridges
the legacy CODER_AGENT_EXP_*_DIRS env vars), moves into
agent.go next to its sole caller.
